+++ test/runtests.jl
@@ -5,7 +5,7 @@ using Arblib, Test, LinearAlgebra, SpecialFunctions
 @testset "Arblib" begin
     include("arb_types-test.jl")
     include("types-test.jl")
-    include("arbcall-test.jl")
+    include("ArbCall-test.jl")
     include("precision-test.jl")
     include("setters-test.jl")
     include("constructors-test.jl")
